パンジャンステージのバグを少し修正
