+++ Assets/Project/Program/PanjanDrum/PanjanExplode.cs
@@ -7,7 +7,7 @@ public class PanjanExplode : MonoBehaviour
 {
     GameManager gameManager;
     bool isEnd;
-    private string ResultSceneName = "Result";
+    PanjanMake panjanMake;
     //GameObject PrefabFire;
     // Start is called before the first frame update
     void Start()
@@ -15,15 +15,10 @@ public class PanjanExplode : MonoBehaviour
         //PrefabFire = Resources.Load<GameObject>("Assets/Design_/Sugi'sTemp/Ninja/Sources/Fire");
         //gameManager取得
         gameManager = GameObject.FindGameObjectWithTag("GameManager").GetComponent<GameManager>();
+        panjanMake=GameObject.Find("PanjanMaker").GetComponent<PanjanMake>();
         //transform=new Vector3(Random.Range(-100, 100), Random.Range(-100, 100), Random.Range(-100, 100));
     }
 
-
-    // Update is called once per frame
-    void Update()
-    {
-        
-    }
 
     private void OnCollisionEnter(Collision other)
     {
@@ -34,35 +29,11 @@ public class PanjanExplode : MonoBehaviour
             {
                 other.transform.Find("Fire").gameObject.SetActive(true);
                 other.transform.GetComponent<Rigidbody>();
-                //SceneManagerのイベントに勝利リザルト処理を追加
-                isEnd = true;
-                //GetComponent<AudioSource>().Play();
-                SceneManager.sceneLoaded += GameSceneLoaded;
-                gameManager.StateChangeSignal(GameState.Finish);
-                Time.timeScale = 0.2f;
-
-                //リザルトへのシーン遷移
-                StartCoroutine(GoResult());
-
+                panjanMake.Burn();
             }
         }
     }
 
-    //スロー演出からのシーン遷移
-    IEnumerator GoResult()
-    {
-        yield return new WaitForSeconds(0.8f);
-        Time.timeScale = 1;
-        SceneManager.LoadScene(ResultSceneName);
-    }
-
-    //パンジャンリザルト用の処理
-    void GameSceneLoaded(Scene next, LoadSceneMode mode)
-    {
-        ResultManager resultManager = GameObject.Find("ResultManager").GetComponent<ResultManager>();
-        resultManager.SetResult(Result.Lose, "爆破");
-
-        SceneManager.sceneLoaded -= GameSceneLoaded;
-    }
+   
 
 }
